0.0.7: Add valid session check in acquireLeadership step
We were seeing cases where our sessions became invalid and we could never
acquire leadership. If we try to acquire leadership and don't get it, check
our session validity before deciding we've lost the election. If session is
no longer valid, reset it and try again.

We may still not get leadership on the first pass thanks to the lock delay,
but we should get the situation sorted out within 30 seconds.